git-svn-id: https://svn.kuali.org/repos/student/trunk@17731 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core/ks-core-ui/src/main/java/org/kuali/student/core/organization/assembly/OrgPersonRelationAssembler.java
@@ -29,13 +29,14 @@ import java.util.List;
 import org.apache.log4j.Logger;
 import org.kuali.student.common.ui.client.mvc.DataModel;
 import org.kuali.student.common.ui.client.mvc.DataModelDefinition;
-import org.kuali.student.core.assembly.Assembler;
 import org.kuali.student.core.assembly.data.AssemblyException;
 import org.kuali.student.core.assembly.data.Data;
 import org.kuali.student.core.assembly.data.Metadata;
 import org.kuali.student.core.assembly.data.QueryPath;
-import org.kuali.student.core.assembly.data.SaveResult;
 import org.kuali.student.core.assembly.data.Data.Property;
+import org.kuali.student.core.assembly.old.Assembler;
+import org.kuali.student.core.assembly.old.data.SaveResult;
+import org.kuali.student.core.assembly.util.AssemblerUtils;
 import org.kuali.student.core.dto.MetaInfo;
 import org.kuali.student.core.exceptions.DoesNotExistException;
 import org.kuali.student.core.organization.assembly.data.server.org.OrgHelper;
@@ -133,9 +134,10 @@ public class OrgPersonRelationAssembler implements Assembler<Data, OrgPersonHelp
                     } catch (Exception e) {
                         throw new AssemblyException();
                     }
+                    AssemblerUtils.setUpdated(orgPersonHelper.getData(), false);
                 }
             }
-            else if(isDeleted(orgPersonHelper.getData())){
+            else if(isDeleted(orgPersonHelper.getData())&&orgPersonHelper.getId()!=null){
                 try{
                     orgService.removeOrgPersonRelation(orgPersonHelper.getId());
                     propertyIter.remove();
@@ -157,6 +159,7 @@ public class OrgPersonRelationAssembler implements Assembler<Data, OrgPersonHelp
                 	LOG.error(e);
                     throw new AssemblyException();
                 }
+                AssemblerUtils.setCreated(orgPersonHelper.getData(), false);
             }
